% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aSPU.R
\name{aSPU}
\alias{aSPU}
\title{Sum of Powered Score (SPU) tests and adaptive SPU (aSPU) test.}
\usage{
aSPU(Y, X, cov = NULL, resample = c("perm", "sim", "boot"),
  model = c("gaussian", "binomial"), pow = c(1:8, Inf), n.perm = 1000)
}
\arguments{
\item{Y}{Response or phenotype data. It can be disease indicators; =0 for controls, =1 for cases.
Or it can be a quantitative trait. A vector with length n (number of observations)}

\item{X}{Genotype or other data; each row for a subject, and each column
    for an SNP (or a predictor). The value of each SNP is the # of the copies
    for an allele. A matrix with dimension n by p (n : number of observation, p : number of genotype data)}

\item{cov}{Covariates. A matrix with dimension n by k (n :number of observation, k : number of covariates)}

\item{resample}{Use "perm" for residual permutations, "sim" for simulations from the null distribution, and "boot" for parametric bootstrap.}

\item{model}{Use "gaussian" for quantitative trait, and use "binomial" for binary trait.}

\item{pow}{power used in SPU test. A vector of the powers.}

\item{n.perm}{number of permutations or bootstraps.}
}
\value{
A list object, Ts : Test Statistics for the SPU and aSPU test.
        pvs : p-values for the SPU and aSPU test.
}
\description{
It gives p-values of the SPU tests and aSPU test.
}
\examples{
data(exdat)

## example analysis using aSPU test on exdat data.
out <- aSPU(exdat$Y, exdat$X, cov = NULL, resample = "boot",
           model = "binomial", pow = c(1:8, Inf), n.perm = 1000)

out$Ts
# These are list of Test Statistics for SPU and aSPU tests.
out$pvs
# These are p-values of Test Statistics for SPU and aSPU tests.
}
\author{
Il-Youp Kwak, Junghi Kim, Yiwei Zhang and Wei Pan
}
\references{
Wei Pan, Junghi Kim, Yiwei Zhang, Xiaotong Shen and Peng Wei (2014)
A powerful and adaptive association test for rare variants,
Genetics, 197(4), 1081-95
}
\seealso{
\code{\link{aSPUw}}
}

