% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estcov.R
\name{estcov}
\alias{estcov}
\title{estcov}
\usage{
estcov(allZ)
}
\arguments{
\item{allZ}{matrix of summary Z-scores for all SNP. each row for SNP; each column for single trait}
}
\value{
estimated correlation matrix
}
\description{
Estimate covariance matrix based on the summary Z-scores
}
\examples{

# -- n.snp: number of SNPs
# -- n.trait: number of traits
# -- n.subject: number of subjects

n.snp <- 100
n.traits <- 10
n.subjects <- 1000
traits <- matrix(rnorm(n.subjects*n.traits), n.subjects, n.traits)
v <- cov(traits)
allZ <- rmvnorm(n.snp, sigma=v)
colnames(allZ) <- paste("trait", 1:n.traits, sep="")
rownames(allZ) <- paste("snp", 1:n.snp, sep="")


v <- estcov(allZ)
MTaSPUs(Z = allZ, v = v, B = 100, pow = c(1:4, Inf), tranform = FALSE)
MTaSPUs(Z = allZ[1,], v = v, B = 100, pow = c(1:4, Inf), tranform = FALSE)
minP(Zi= allZ[1,], v = v)

}
\author{
Junghi Kim, Yun Bai and Wei Pan
}
\references{
Junghi Kim, Yun Bai and Wei Pan (2016) An Adaptive Association Test for Multiple Phenotypes with GWAS Summary Statistics, Genetic Epidemiology DOI 10.1002/gepi.21931
}
\seealso{
\code{\link{MTaSPUs}} \code{\link{minP}}
}

