% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GATES2.R
\name{GATES2}
\alias{GATES2}
\title{Gates}
\usage{
GATES2(ldmatrix, p)
}
\arguments{
\item{ldmatrix}{The correlation matrix among SNPs,}

\item{p}{p-value for each SNPs.}
}
\value{
Combined p-values using GATES method and the key gene location.
}
\description{
Get p-value using GATES method. Usually it used to get genomewise p-values.
This function is taken from postgwas package.
There is a little modification of the code GATES in postgwas package
to get key gene location for Hyst method.
}
\examples{
simula <- simPathAR1Snp(nGenes=20, nGenes1=1, nSNPlim=c(1, 20), nSNP0=1:3,
                           LOR=.2, rholim=c(0,0),
                           n=100, MAFlim=c(0.05, 0.4), p0=0.05)
logitp <- getlogitp(simula$Y, simula$X)

## get correlation of SNPs using controls
ldmat <- cor(simula$X[ simula$Y == 0, ])
(gatesp <- GATES2(ldmat, logitp))
}
\author{
Milan Hiersche(taken from pastgwas package), Il-Youp Kwak(modified a little)
}
\references{
Miao-Xin Li, Hong-Sheng Gui, Johnny S.H. Kwan and Pak C. Sham (2011)
GATES: A Rapid and Powerful Gene-Based Association Test Using Extended Simes Procedure
}
\seealso{
\code{\link{Hyst}} \code{\link{GatesSimes}}
}

