% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEEaSPU.R
\name{GEEaSPU}
\alias{GEEaSPU}
\title{GEEaSPU test.}
\usage{
GEEaSPU(traits, geno, Z = NULL, family = c("binomial", "gaussian"),
  gamma = c(1:8, Inf), n.perm = 10000)
}
\arguments{
\item{traits}{trait matrix. The row for individuals and the column for traits.}

\item{geno}{A matrix of genetic information.}

\item{Z}{covariates.}

\item{n.perm}{number of permutations or bootstraps.}

\item{pow}{power used in GEEaSPU test. A vector of the powers.}
}
\value{
p-values for the SPUs and aSPUs tests.
}
\description{
It gives p-values of the GEESPU tests and GEEaSPU test.
}
\examples{
traits <- matrix(rnorm(100*5, 0,1), ncol=5)
Z <- rnorm(100, 2, 0.5)
geno <- rbinom(100, 2, 0.5)
out <- GEEaSPU(traits, geno, Z = NULL,family = "gaussian",
		  gamma = c(1:8,Inf), n.perm = 1000)
}
\author{
Il-Youp Kwak and Wei Pan
}
\references{
Il-Youp Kwak, Wei Pan (2015)
Adaptive Gene- and Pathway-Trait Association Testing with GWAS Summary Statistics,
in revision.

Wei Pan, Junghi Kim, Yiwei Zhang, Xiaotong Shen and Peng Wei (2014)
A powerful and adaptive association test for rare variants,
Genetics, 197(4), 1081-95

Junghi Kim, Jeffrey R Wozniak, Bryon A Mueller, Xiaotong Shen and Wei Pan (2014) Comparison of statistical tests for group differences in brain functional networks, NeuroImage, 1;101:681-694
}

