% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minP.R
\name{minP}
\alias{minP}
\title{minP test.}
\usage{
minP(Zi, v)
}
\arguments{
\item{Zi}{a vector of summary Z-scores for single SNP}

\item{v}{estimated covariance matrix based on the summary Z-scores (output of estcov)}

\item{tranform}{if TRUE, the inference is made on transformed Z}

\item{B}{number of Monte Carlo samples simulated to compute p-values}
}
\value{
return exact minP test
}
\description{
Return exact minP test.
}
\examples{

# -- n.snp: number of SNPs
# -- n.trait: number of traits
# -- n.subject: number of subjects

n.snp <- 100
n.traits <- 10
n.subjects <- 1000
traits <- matrix(rnorm(n.subjects*n.traits), n.subjects, n.traits)
v <- cov(traits)
allZ <- rmvnorm(n.snp, sigma=v)
colnames(allZ) <- paste("trait", 1:n.traits, sep="")
rownames(allZ) <- paste("snp", 1:n.snp, sep="")


v <- estcov(allZ)
sumaspu(Z = allZ, v = v, B = 100, pow = c(1:4, Inf), tranform = FALSE)
sumaspu(Z = allZ[1,], v = v, B = 100, pow = c(1:4, Inf), tranform = FALSE)
minP(Zi= allZ[1,], v = v)

}
\author{
Junghi Kim, Yun Bai and Wei Pan
}
\references{
Junghi Kim, Yun Bai and Wei Pan (2016) An Adaptive Association Test for Multiple Phenotypes with GWAS Summary Statistics, Genetic Epidemiology DOI 10.1002/gepi.21931
}
\seealso{
\code{\link{estcov}} \code{\link{sumaspu}}
}

