% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumaspu.R
\name{sumaspu}
\alias{sumaspu}
\title{An Adaptive Association Test for Multiple Phenotypes with GWAS Summary Statistics.(SNP based)}
\usage{
sumaspu(Z, v, pow = c(1:4, Inf), B)
}
\arguments{
\item{Z}{matrix of summary Z-scores, SNPs in rows and traits in columns. Or a vector of summary Z-scores for a single snp}

\item{v}{estimated covariance matrix based on the summary Z-scores (output of estcov)}

\item{pow}{power used in SPU test. A vector of the powers.}

\item{B}{number of Monte Carlo samples simulated to compute p-values}

\item{tranform}{if TRUE, the inference is made on transformed Z}
}
\value{
compute p-values for SPU(gamma) i.e. pow=1:8, and infinity
            aSPU, based on the minimum p-values over SPU(power)
            each row for single SNP
}
\description{
SNP based Adaptive Association Test for Multiple Phenotypes with GWAS Summary Statistics.
}
\examples{

# -- n.snp: number of SNPs
# -- n.trait: number of traits
# -- n.subject: number of subjects

n.snp <- 100
n.traits <- 10
n.subjects <- 1000
traits <- matrix(rnorm(n.subjects*n.traits), n.subjects, n.traits)
v <- cov(traits)
allZ <- rmvnorm(n.snp, sigma=v)
colnames(allZ) <- paste("trait", 1:n.traits, sep="")
rownames(allZ) <- paste("snp", 1:n.snp, sep="")


v <- estcov(allZ)
sumaspu(Z = allZ, v = v, B = 100, pow = c(1:4, Inf), tranform = FALSE)
sumaspu(Z = allZ[1,], v = v, B = 100, pow = c(1:4, Inf), tranform = FALSE)
minP(Zi= allZ[1,], v = v)

}
\author{
Junghi Kim, Yun Bai and Wei Pan
}
\references{
Junghi Kim, Yun Bai and Wei Pan (2016) An Adaptive Association Test for Multiple Phenotypes with GWAS Summary Statistics, Genetic Epidemiology DOI 10.1002/gepi.21931
}
\seealso{
\code{\link{minP}} \code{\link{estcov}}
}

