% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEEaSPU.R
\name{GEEaSPU}
\alias{GEEaSPU}
\title{GEEaSPU test for association with multiple traits in generalized estimation equations.}
\usage{
GEEaSPU(traits, geno, Z = NULL, model = c("binomial", "gaussian"),
  gamma = c(1:8, Inf), n.sim = 1000)
}
\arguments{
\item{traits}{trait matrix. The row for individuals and the column for traits.}

\item{geno}{A matrix of genetic information.}

\item{Z}{covariates.}

\item{model}{Use "gaussian" for a quantitative trait, and use "binomial" for a binary trait.}

\item{gamma}{power used in GEEaSPU test. A vector of the powers.}

\item{n.sim}{number of simulations}
}
\value{
p-values for the GEE-SPU and GEE-aSPU test.
}
\description{
It gives p-values of the GEESPU tests and GEEaSPU test.
}
\examples{
traits <- matrix(rnorm(100*5, 0,1), ncol=5)
Z <- rnorm(100, 2, 0.5)
geno <- rbinom(100, 2, 0.5)
out <- GEEaSPU(traits, geno, Z = NULL, model = "gaussian",
		  gamma = c(1:8,Inf), n.sim = 100)
}
\author{
Junghi Kim, Wei Pan and Il-Youp Kwak
}
\references{
Yiwei Zhang, Zhiyuan Xu, Xiaotong Shen, Wei Pan (2014)
Testing for association with multiple traits in generalized estimation equations, with application to neuroimaging data.
Neuroimage. 1;96:309-25
}

