% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Hyst.R
\name{Hyst}
\alias{Hyst}
\title{HYST}
\usage{
Hyst(pvec, ldmatrix, snp.info, gene.info)
}
\arguments{
\item{pvec}{p-values for each SNP.}

\item{ldmatrix}{numeric. A correlation matrix of SNPs, dimensions matching the p and snps arguments.}

\item{snp.info}{A matrix with snp data. Data are coded as 0, 1, 2,
corresponding to homozygotes for the major allele,
heterozygotes, and homozygotes for the minor allele,
respectively. Each row of the matrix is one SNP and each
column represents one subject.}

\item{gene.info}{A matrix with four columns with the order of gene.Name,
chr, Start, and End. The name of the gene is under gene.Name.
The chromosome is chr, and the start and end coordinate
(start < end) are in the Start and End columns. It is suggested
that the genes are sorted by their locations in the genome.}
}
\value{
p-values for each SNPs.
}
\description{
Get p-value using GATES-Simes method.
}
\examples{
simula <- simPathAR1Snp(nGenes=20, nGenes1=1, nSNPlim=c(1, 20), nSNP0=1:3,
                           LOR=.2, rholim=c(0,0),
                           n=100, MAFlim=c(0.05, 0.4), p0=0.05)
logitp <- getlogitp(simula$Y, simula$X)

## get correlation of SNPs using controls
ldmat <- cor(simula$X[ simula$Y == 0, ])
out <- Hyst(pvec = logitp, ldmatrix = ldmat, snp.info = simula$snp.info,
            gene.info = simula$gene.info)
}
\author{
Il-Youp Kwak
}
\references{
Miao-Xin Li, Johnny S.H. Kwan and Pak C. Sham (2012)
HYST: A Hybrid Set-Based Test for Genome-wide Association Studies, with Application to Protein-Protein Interaction-Based Association Analysis
}
\seealso{
\code{\link{aSPUw}}
}

